requirements-dev.txt
coreapi
django
django-extensions
djangorestframework
djangorestframework-jwt
gunicorn
locustio
pdbpp
pytest
pytest-django